+++ app/src/main/java/com/santimattius/template/MainApplication.kt
@@ -1,5 +1,7 @@
 package com.santimattius.template
 
 import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
 
+@HiltAndroidApp
 class MainApplication : Application()
