+++ app/src/main/java/com/santimattius/template/domain/repositories/MovieRepository.kt
@@ -3,8 +3,6 @@ package com.santimattius.template.domain.repositories
 import com.santimattius.template.domain.entities.Movie
 
 interface MovieRepository {
-
     suspend fun getPopular(): List<Movie>
-
     suspend fun fetchPopular(): Result<List<Movie>>
 }
