+++ app/src/main/java/com/santimattius/template/ui/home/MainActivity.kt
@@ -3,7 +3,9 @@ package com.santimattius.template.ui.home
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.santimattius.template.R
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
