+++ gradle/libs.versions.toml
@@ -2,23 +2,21 @@
 # Plugins
 androidGradlePlugin = "8.0.0"
 kotlinGradlePlugin = "1.8.10"
-hiltGradlePlugin = "2.43.1"
+hiltGradlePlugin = "2.44"
 detektGradlePlugin = "1.22.0"
 # Define the dependency versions
-kotlin = "1.7.10"
-compose = "1.3.1"
-activityCompose = "1.6.1"
+kotlin = "1.8.10"
 coreKtx = "1.9.0"
 coreTesting = "2.1.0"
 legacySupport = "1.0.0"
 appCompat = "1.4.1"
-fragmentKtx = "1.4.1"
-fragmentTesting = "1.4.1"
+fragmentKtx = "1.5.2"
+fragmentTesting = "1.5.2"
 constraintLayout = "2.1.3"
 recyclerView = "1.2.1"
 materialVersion = "1.5.0"
 room = "2.5.0"
-hilt = "2.43.1"
+hilt = "2.44"
 coil = "2.1.0"
 
 junit = "4.13.2"
